shoulder_mri/requirements.txt
Flask
Flask-SQLAlchemy
Pillow
tensorflow==2.15.1
numpy
